+++ spec/gyro/liquid_generator_java_spec.rb
@@ -20,7 +20,7 @@ module Gyro
   describe 'Liquid' do
     describe 'Java' do
       before do
-        Gyro::Log.quiet = true
+        Gyro::Log.in_test_context = true
       end
       JAVA_MODELS.each do |datamodel|
         it datamodel do