+++ spec/gyro/liquid_generator_kotlin_spec.rb
@@ -23,7 +23,7 @@ module Gyro
   describe 'Liquid' do
     describe 'Kotlin' do
       before do
-        Gyro::Log.quiet = true
+        Gyro::Log.in_test_context = true
       end
 
       KOTLIN_MODELS.each do |datamodel|