+++ CHANGELOG.md
@@ -6,6 +6,13 @@
   _Now there is a config.yml inside template directory to manage alias and deprecated._  
   [Steven Watremez](https://github.com/StevenWatremez)
   [#35](https://github.com/NijiDigital/gyro/pull/35)
+* Fix entity comments for swift templates. Improve comments on attributes. Add comments on relationships.  
+  _Now you can use comment  info key on entity, attribute and relationship_.  
+  [Steven Watremez](https://github.com/StevenWatremez)
+  [#56](https://github.com/NijiDigital/gyro/pull/56)
+* Improve error and warning messages when targetting an xcdatamodeld.  
+  [Steven Watremez](https://github.com/StevenWatremez)
+  [#53](https://github.com/NijiDigital/gyro/pull/53)
 
 ## 1.4.0